refactor: improve api behaviour

+++ grader.sh
@@ -48,11 +48,9 @@ check_password_file_contains() {
     grep -q "$1" /var/secure/user_passwords.txt || grep -q "$1" /var/secure/user_passwords.csv
 }
 
-echo "[" >$RESULT_FILE
-
 # Check if the repository exists and is public
 if ! curl -s -o /dev/null -w "%{http_code}" "$REPO_URL" | grep -q "200"; then
-    echo "GitHub repository does not exist or is not public" >$RESULT_FILE
+    echo "{\"score\": 0, \"error\": \"GitHub repository does not exist or is not public\"}" >$RESULT_FILE
     exit 1
 fi
 
@@ -60,44 +58,44 @@ cleanup
 git clone $REPO_URL $CLONE_DIR
 cd $CLONE_DIR
 
+# Check if the create_users.sh script exists
+if [ ! -f create_users.sh ]; then
+    echo "{\"score\": 0, \"error\": \"create_users.sh script does not exist in the repository\"}" >$RESULT_FILE
+    exit 1
+fi
+
 tests=(
-    "Test 1: Check if script exists;\
-    [ -f create_users.sh ]"
-
-    "Test 2: Check if README.md exists;\
-    [ -f README.md ]"
-
-    "Test 3: Check if users are created;\
+    "Test 1: Check if users are created;\
     echo -e \"zxenon; sudo,hng,www-data\nidimma; sudo\nmayowa; hng,www-data\" > userlist.txt &&\
     sudo bash create_users.sh userlist.txt &&\
     check_user_exists \"zxenon\" &&\
     check_user_exists \"idimma\" &&\
     check_user_exists \"mayowa\""
 
-    "Test 4: Check if groups are created;\
+    "Test 2: Check if groups are created;\
     check_group_exists \"hng\""
 
-    "Test 5: Check if personal group is created for each user;\
+    "Test 3: Check if personal group is created for each user;\
     check_group_exists \"zxenon\" &&\
     check_group_exists \"mayowa\""
 
-    "Test 6: Check if users belong to all specified groups;\
+    "Test 4: Check if users belong to all specified groups;\
     check_user_in_group \"zxenon\" \"sudo\" &&\
     check_user_in_group \"mayowa\" \"hng\" &&\
     check_user_in_group \"zxenon\" \"www-data\""
 
-    "Test 7: Check if users belong to their personal group;\
+    "Test 5: Check if users belong to their personal group;\
     check_user_in_group \"zxenon\" \"zxenon\" &&\
     check_user_in_group \"mayowa\" \"mayowa\""
 
-    "Test 8: Check if passwords are stored securely;\
+    "Test 6: Check if passwords are stored securely;\
     check_password_file_contains \"zxenon\" &&\
     check_password_file_contains \"mayowa\""
 
-    "Test 9: Check if logs are stored securely;\
+    "Test 7: Check if logs are stored securely;\
     check_log_contains \"zxenon\""
 
-    "Test 10: Check if groups are created and user is added even if user already exists;\
+    "Test 8: Check if groups are created and user is added even if user already exists;\
     sudo useradd -m -s /bin/bash testuser &&\
     echo \"testuser; testgroup\" > userlist.txt &&\
     sudo bash create_users.sh userlist.txt &&\
@@ -105,18 +103,26 @@ tests=(
     check_user_in_group \"testuser\" \"testgroup\""
 )
 
+total_tests=${#tests[@]}
+passed_tests=0
+
+results=()
+
 for i in "${!tests[@]}"; do
     test="${tests[i]}"
     description="${test%%;*}"
     command="${test#*; }"
     echo "Running $description"
-    eval "$command" && {
-        echo "{\"title\": \"$description\", \"status\": \"pass\"}" >>$RESULT_FILE
-    } || {
-        echo "{\"title\": \"$description\", \"status\": \"fail\"}" >>$RESULT_FILE
-    }
-    [ $i -lt $((${#tests[@]} - 1)) ] && echo "," >>$RESULT_FILE
+    if eval "$command"; then
+        results+=("{\"title\": \"$description\", \"status\": \"pass\"}")
+        passed_tests=$(expr $passed_tests + 1)
+    else
+        results+=("{\"title\": \"$description\", \"status\": \"fail\"}")
+    fi
 done
 cleanup
 
-echo "]" >>$RESULT_FILE
+results_json=$(IFS=,; echo "${results[*]}")
+
+echo "{\"score\": $passed_tests, \"result\": [$results_json]}" >$RESULT_FILE
+
